refactor: rename _common.py -> common.py, _base.py -> base.py (#2907)
